move validate to ui-common
