+++ packages/accounts-app/package.json
@@ -13,7 +13,6 @@
     "test": "echo skipped"
   },
   "dependencies": {
-    "@substrate/transfer-app": "^0.2.0",
     "@substrate/ui-common": "^0.2.0",
     "@substrate/ui-components": "^0.2.0",
     "fp-ts": "^1.16.1"
