+++ packages/accounts-app/src/Accounts/Staking/Bond.tsx
@@ -6,8 +6,7 @@ import { SubmittableExtrinsic } from '@polkadot/api/SubmittableExtrinsic';
 import { DerivedFees, DerivedBalances } from '@polkadot/api-derive/types';
 import { Index, getTypeRegistry } from '@polkadot/types';
 import { isUndefined } from '@polkadot/util';
-import { validate } from '../../validate';
-import { AppContext, AlertsContext, TxQueueContext } from '@substrate/ui-common';
+import { AppContext, AlertsContext, TxQueueContext, validate } from '@substrate/ui-common';
 import { Dropdown, DropdownProps, Input, Stacked, SubHeader, WithSpaceAround, StyledNavButton } from '@substrate/ui-components';
 import BN from 'bn.js';
 import React, { useContext, useState, useEffect } from 'react';
