+++ packages/ui-common/src/util/index.ts
@@ -5,3 +5,4 @@
 export * from './isTestChain';
 export * from './keyring';
 export * from './ui';
+export * from './validate';
